+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/curse/CurseOfWalking.java
@@ -1,0 +1,31 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.curse;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class CurseOfWalking extends Enchantment {
+    public CurseOfWalking() {
+        super(Rarity.RARE, EnchantmentCategory.ARMOR_LEGS, new EquipmentSlot[]{EquipmentSlot.LEGS});
+    }
+
+    @Override
+    public boolean isTreasureOnly() {
+        return true;
+    }
+
+    @Override
+    public boolean isCurse() {
+        return true;
+    }
+
+    public int getMinCost(int pEnchantmentLevel) {
+        return 25;
+    }
+
+    public int getMaxCost(int pEnchantmentLevel) {
+        return 50;
+    }
+
+
+}