+++ changelog.md
@@ -7,6 +7,8 @@
   * A new pants treasure enchantment that greatly increases movement speed at the cost of not being able to run anymore
 * Added Curse of Slow charge
   * Opposite of Quick Charge
+* Added Curse of Walking
+  * You can no longer walk
 * Reworked M.A. Jump
   * Now called Double Jump, it's one level only and is a treasure enchantment
   * Any equipment with it should be disenchanted as it has an invalid enchantment now